requirements.txt
eve==0.8
gunicorn==19.9.0
pyjwt==1.7.1
python-ldap==3.1.0